Shim a new base class for drawing-like tools
The existing `AbstractDrawingTool` is now in a transitional state and will not exist much longer. The new-style tools
will not use it, but will inherit from `AbstractDrawingLikeTool`, which provides a bare minimum of
functionality. Eventually, `AbstractDrawinTool` will be repurposed into a base class for templates, and even then may
not look much like it does now.
